LP-83 Add token popover
